Refactor game slice types
